refactored all event consumer interfaces into a separate platform-agnostic file
